Add EndpointHandler, EndpointErrorHandler and EndpointMiddleware into public API - Divide transaction types into own module
