+++ src/index.ts
@@ -5,6 +5,7 @@ export { HorsServer } from './server';
 export { Endpoint } from './endpoint';
 
 export { Transaction, TransactionRequestInfo } from './transaction/transaction';
+export { EndpointHandler, EndpointMiddleware, EndpointErrorHandler } from './transaction/transaction-types';
 
 export { apiEndpoint, ApiHttpMethod } from './decorators/api-endpoint';
