+++ src/transaction/transaction-utils.ts
@@ -1,14 +1,6 @@
-import { Request, RequestHandler, ErrorRequestHandler, NextFunction } from 'express';
+import { RequestHandler, ErrorRequestHandler } from 'express';
 
-import Transaction from './transaction';
-
-export type EndpointHandler = (transaction: Transaction) => void;
-export type EndpointErrorHandler = (transaction: Transaction, error: any) => void;
-export type EndpointMiddleware = (transaction: Transaction, next: NextFunction) => void;
-
-export interface RequestWithTransaction extends Request {
-    transaction: Transaction;
-}
+import { EndpointHandler, EndpointErrorHandler, EndpointMiddleware, RequestWithTransaction } from './transaction-types';
 
 /**
  * Converts a transaction based endpoint handler into Express endpoint handler
